refactor(intelligence-daemon): decompose 1116-line daemon into 3 focused modules
Extract PRDiscovery, DashboardBuilder, HygieneRunner from IntelligenceDaemon.
Daemon reduces from 1116 to 499 lines as a thin orchestrator. All 83 existing
tests pass. No behavior change — pure structural refactor.

Dispatch-ID: 20260408-160000-intelligence-daemon-decompose-A
